stocktrack/requirements.txt
django
djangorestframework
markdown
django-filter
psycopg2
psycopg2-binary